LastError to own header + doxygen
